refactor: PurchasedLottoTicketsFactory가 RandomLottoTicketFactory를 주입받도록 변경
