+++ src/main/java/lotto/domain/PurchasedLottoTicketsFactory.java
@@ -5,12 +5,12 @@ import java.util.List;
 
 public class PurchasedLottoTicketsFactory {
 	public static PurchasedLottoTickets of(PurchaseMoney purchaseMoney,
-										   RandomGenerator randomGenerator) {
+										   RandomLottoTicketFactory randomLottoTicketFactory) {
 		List<SerialLottoNumber> purchasedLottoTickets = new ArrayList<>();
 
 		for (int i = 0; i < purchaseMoney.countPurchasedTickets(); i++) {
 			purchasedLottoTickets.add(
-					SerialLottoNumberFactory.createRandomLottoTicket(randomGenerator));
+					randomLottoTicketFactory.create());
 		}
 
 		return new PurchasedLottoTickets(purchasedLottoTickets);
