+++ src/main/java/lotto/domain/WinningLottoNumbersFactory.java
@@ -1,17 +1,11 @@
 package lotto.domain;
 
-import lotto.exceptions.LottoTicketIllegalArgumentException;
 import lotto.utils.StringParser;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class SerialLottoNumberFactory {
-	public static SerialLottoNumber createRandomLottoTicket(RandomGenerator randomGenerator)
-			throws LottoTicketIllegalArgumentException {
-		return new SerialLottoNumber(randomGenerator.generateSixNumbers());
-	}
-
+public class WinningLottoNumbersFactory {
 	public static SerialLottoNumber createWinningLottoNumbers(String input)
 			throws IllegalArgumentException {
 		List<Integer> integers = StringParser.stringToIntegerList(input);
