+++ src/test/java/lotto/domain/RandomLottoTicketFactoryTest.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-public class SerialLottoNumberFactoryTest {
+public class RandomLottoTicketFactoryTest {
 	@Test
 	void createRandomLottoTicket() {
 		// given
@@ -15,27 +15,12 @@ public class SerialLottoNumberFactoryTest {
 
 		// when
 		SerialLottoNumber randomLottoTicket =
-				SerialLottoNumberFactory.createRandomLottoTicket(randomLottoNumbersGenerator);
+				new RandomLottoTicketFactory(randomLottoNumbersGenerator).create();
 
 		// then
 		List<LottoNumber> lottoNumbers = Stream.of(1, 2, 3, 4, 5, 6)
 				.map(AllLottoNumbers::getLottoNumber)
 				.collect(Collectors.toList());
 		Assertions.assertThat(randomLottoTicket).isEqualTo(new SerialLottoNumber(lottoNumbers));
-	}
-
-	@Test
-	void createWinningLottoNumbers() {
-		// given
-		String input = "1, 2, 3, 4, 5, 6";
-
-		// when
-		SerialLottoNumber winningLottoNumbers = SerialLottoNumberFactory.createWinningLottoNumbers(input);
-
-		// then
-		List<LottoNumber> lottoNumbers = Stream.of(1, 2, 3, 4, 5, 6)
-				.map(AllLottoNumbers::getLottoNumber)
-				.collect(Collectors.toList());
-		Assertions.assertThat(winningLottoNumbers).isEqualTo(new SerialLottoNumber(lottoNumbers));
 	}
 }
